PYI-465: Give parameters real values
The previous approach of using dummy data for the parameter values which
we'll update outside of terraform won't work. Terraform will just try to
overwrite it the next time it's applied. Setting the `overwrite` flag to
false didn't do what I'd expected it to. Instead it just throws an
error.

This change uses values for the certs passed in from the terraform
project root in `di-ipv-config` for the non-secure parameters. We'll
need to manage the "SecureString" parameters completely outside of
Terraform as we can't pass the secret values in in the same way.