PYI-476: Add stub DCS lambda
This adds a lambda that will only be deployed locally. It's a stub for
DCS.

Currently it ignores the incoming payload and generates a valid response
with new UUIDs.

When we've finished the work in the passport CRI to verify and decrypt
response from the DCS, we can steal the logic to unwrap incoming
payloads to the stub DCS and use the UUIDs contained in the response.

It will use dev PKI that is passed in from the terraform project root in
local-startup. It goes without saying that this PKI should never be used
for anything else.