PYI-476: Verify/decrypt payload in stub DCS
This adds verification and decryption of the incoming payload from the
Passport CRI in the stub DCS. We then return a response to the CRI using
the correlation and request IDs contained in the original request.

This will help give us confidence that the request we're sending to DCS
is legit.