PYI-452 Parse DCS response after unwrapping
After unwrapping the DCS response parse it into a DcsResponse object and
persist this in the database.

The methods now appear in the order they are called from the
handleRequest method.

Try to make the handler function simpler to understand by standardising
the exception handling around a common exception. Each process the
handler calls should return the expected object or throw a
HttpResponseExceptionWithErrorBody which can then be dealt with by the
common catch block.

Co-authored-by Daniel Pomfret
<daniel.pomfret@digital.cabinet-office.gov.uk>

Co-authored-by Dan Worth
<dan.worth@digital.cabinet-office.gov.uk>