PYI-492: Set up an environment for running integration tests
This wraps the intTest gradle task with tasks to start up and tear down
an environment for running integration tests in.

It starts a localstack container and applies specific pieces of
Terraform into it that are needed for integration tests. At the moment
it's just dynamoDB tables.

This is good as it means we're running the tests against infrastructure
created with the same Terraform we're using in prod.

It would be nice to be able to tear it all down and set it up again
between tests, but this won't do that...