PYI-612: Add shared attributes signing cert var
The signing cert for verifying the shared attributes JWS has been added
to the config repo. We need to add this variable to stop the terraform
from breaking.

This sets the cert as a parameter in the parameter store. The name of
the parameter may need changing now we're managing config in a
structured way.