PYI-614: Use new configuration structure
The structure of the config in the parameter store has changed. This
updates the deploy template to reflect this. It also makes a small
change to the ConfigurationService code.

It also removes the deploy of the stub-dcs lambda. This lambda was only
ever supposed to be deployed locally. In the dev environment we can talk
to the integration DCS directly.

If we get a local running solution in the future we can resurrect it, so
i'm leaving the stub DCS code where it is for now.