Setup New Integration Test
Use SAM to create DynamoDb tables for Integration Test use in a separate
stack to our main environments. The integration tests can then be run
against our Dev account (for now) either locally or as part of CI (to be
done at a later date). The integration tests are run separately to the
unit test by running the `gradle intTest` task.

The single test added in this commit deletes the item it adds to the
table at the end of the test. Test should be written to not interfere
with concurrent runs of the integration tests, e.g. partition key values
should be random per test.