BAU Pass DCS Response Table Name to integration test
Passing the DCS Response Table name to the integration test means the
tests are independent to how the table may be provisioned for use by
the tests. For example this permits running the test from a developer
machine against their developer environment, or during CI they can be
run against an ephemeral stack or perhaps a longer lived build
environment.

This commit also removes the template to create an integration stack for
use by this test; which at the time of this commit was only creating a
DCS Response Table. On reflection this approach seems flawed because the
table definitions can diverge between the integration stack and the
golden source of the definition within the "real" SAM template.

The DCS_RESPONSE_TABLE_NAME is checked within the BeforeAll method
because this provides better presentation of the error if it is not set
then if this was set as a field on the test class; in such cases Junit
throws a class initialization error and the underlying problem is
somewhat hidden.