PYIC-531: Run int tests in Github workflows
We want to run the integration tests as part of our Github workflows. To
do this Github actions will need to talk to DynamoDB in the build
account.

We already have a Github OIDC provider deployed in AWS which we use to
put build artifacts into S3. We can use the same patten to access
DynamoDB. A separate PR has been raised on the config dir to give the
AWS role perms to talk to DynamoDB.