PYIC-927: Give issuecredential lambda SQS perms
Add the perms required for the issuecredential lambda to write to the
audit event SQS queue.

`SQSSendMessagePolicy` gives the message write permissions. We also need
the KMS permissions as the queue is configured to encrypt messages at
rest and SAM doesn't have a policy template that covers this (obviously
required) use case.

This also adds the SQS queue URL to the lambda as an env var so the lib
actually sending the messages know where to send them.

Whilst not strictly in scope for the core team for slice 2, I've added
it to this lambda first as the KBV teams designated slice 2 message is
`KBV_ADDRESS_CREDENTIAL_ISSUED`. So assuming a matching first message
from the Passport CRI, this lambda is the correct place.