PYIC-1066: Lambda and API gateway logs to Splunk
Add CloudFormation resources to start sending our lambda application
logs and API gateway access logs to Splunk.

For the API gateway logs I've reused the JSON format we have defined for
the core-front Fargate API gateway logs, for consistency. The only
difference is the addition of responseLatency which could be helpful
when looking at slow lambdas.

For the lambdas I've explicitly declared a log group resource even
though lambda creates one for you. This is because the lambda created
log groups will store messages indefinitely by default. You can change
this by declaring the log group yourself. As long as the name is the
same as the log group lambda creates and writes to, the function should
write to this new log group. It will be interesting to see if these log
groups deploy cleanly, considering they already exist...

The subscription filters are taking the logs from the newly created log
groups and sending them to Cyber's Centralised Security Logging Service
(CSLS) Kinesis data stream. This is the recommended way of getting our
logs to Splunk. See here for more details:
https://github.com/alphagov/centralised-security-logging-service