PYIC-1038: Return VC that conforms to RFC0024
Update the VC returned by passport CRI to match the spec defined in
RFC0024:

https://github.com/alphagov/digital-identity-architecture/blob/main/rfc/0024-identity-vc-for-checks.md

The structure defined requires much less data than what we were
previously putting in it. This also makes a few other changes of note.

The `@context` attribute of the credential subject is no longer
required.

The credential subject now includes a passport attribute which contains
the passport number and expiry. This is modeled with a domain object.

The evidence object now contains a `txn` and `type` field.

The `validFrom` and `validUntil` attributes of the credential subject name
parts are not required and have been removed.

The DcsResponse is no longer stored in the DcsPayload. We only really
need the true/false value returned by DCS to generate a GPG45 score. We
don't care about the other data. We will include a contra-indicator in
future however.

The PassportCheckDao `gpg45Score` field has been renamed to `evidence`
to remain consistent with the data model.