PYIC-1038: Set txn to String type
The Nimbus lib uses it's own json stringifier which struggles to
correctly deserialize a UUID. It converts it to an empty JSON object,
rather than a string representation of the UUID. For some reason.

This causes the core to get upset when trying to deserilize an issued
credential.

Passing in the txn this way also makes for looser coupling and
potentially easier testing.