PYIC-1171 Add Provisioned Capacity Parameter
This adds a parameter named ProvisionedConcurrentExecutions to the CFN
template. When this is set to a value greater than 0 each Passport Back
lambda will be provisioned that number of concurrent executions as a
means to reduce the impact of cold starts. The default value is 0.

This will permit turning on and adjusting provisioned concurrent
executions via CloudFormation update-stack operations.

In order to enable provisioned concurrent executions it is necessary to
also enable and create an alias which has been named 'live'.