PYIC-1238: Return `aud` claim in VC JWT
The core is now validating the claims set returned to it from a CRI.
This includes checking that the audience claim has the expected value.

This updates passport to return that claim with a value from the clients
configuration.

This is currently piggybacking on the clients issuer value from config.
At the moment these (issuer and audience) are the same with core. But we
need to look more closely at how we are managing these values as our
current approach is quite confusing.