BAU: Fix KMS signature padding issue
Use nimbus's Base64URL.encode() function to base64URL encode the
signature bytes received from KMS.

The encoder offered by java.util via `Base64.getUrlEncoder()` does not
strip padding as we expected. It's java docs reference Table 2 of RFC
4648, which is a bit vague about the handling of padding:

"This encoding is technically identical to the previous one, except
   for the 62:nd and 63:rd alphabet character, as indicated in Table 2."

As a result we were occasionally including an `=` on the end of our
serialized JWT strings. The nimbus lib is gracious enough to handle
these and still verify the signatures correctly, but it was causing
issues in SPOT. So here's a fix.