BAU: Define common versions for libs
We were using different versions of the same libs in different lambdas.

This defines a structure in the root project where we can define the
versions we want to use and then pull them in in the sub projects.

This makes it easy to bump all versions in case of a security issue, as
well as ensuring consistency to help diagnose weird bugs.

This also sorts all the dependencies to scratch an itch.