BAU: Use SAM to create API key
When trying to deploy the API gateway and the API usage plan at the same
time, cloudformation throws an error. The usage plan can't create as it
can't find the API stage. There is some suggestion in a GitHub issue on
the AWS SAM repo that there is a race condition:

https://github.com/aws/serverless-application-model/issues/1560

The suggested fix is to ensure that you're creating your usage plan with
SAM, rather than with cloudformation as we're currently doing. By
specifying the `UsagePlan` property of an `AWS::Serverless::Api` type,
SAM generates cloudformation resources for `AWS::ApiGateway::UsagePlan`,
`AWS::ApiGateway::ApiKey` and `AWS::ApiGateway::UsagePlanKey`. This
means we can remove them from the template. See the docs:

https://docs.aws.amazon.com/serverless-application-model/latest/developerguide/sam-property-api-apiusageplan.html

I'm guessing this change will change the API key currently being used,
so we will have to go unbreak some stuff.