PYCI-1393: Remove token auth method configuration
Our OAuth spec states that a CRI should enforce the private_key_jwt
client authentication method. We currently have the ability to bypass
this with some config for client. We're using private keys for auth in
all envs and are well practiced at it. The configuration was originally
added to help us build out this feature. As we don't really need this
functionality anymore we should remove this potential foot cannon.