BAU: Throw SsmException instead of ParameterNotFoundException
We validate the client ID received in the auth request params by trying
to fetch some arbitary config for that client from the param store. If
we are able to we consider the client ID validated.

If we don't have any configuration for that client ID the powertools lib
will throw. We were expecting to catch a ParameterNotFound exception,
however that's not what we get. Instead we get a SsmException due to the
lambda not having permissions to get that parameter (even though it
doesnt' exist).

ParameterNotFoundException actually extends SsmException, so if for some
reason the lambda DID have perms and threw that specific exception we'd
still catch it here.