BAU: Fix JSON logging
The logging was broken with the move to using log4j2 for logging in JSON
format. This was due to a missing dependency that allows parsing of YAML
files. log4j2 was failing to parse the YAML configuration file and
falling back to its default configuration. The default config only logs
warn and above.