PYIC-1488: Add IDs to JSON logs
This updates each lambda to add the passport session ID, the client ID
and a component ID to the logs as soon as they're available.

The component ID is a constant thing we add to help with filtering logs.
The same value will be added in the frontend logs. I've intentionally
left it environment agnostic (so not a URL) so it's easy to handle
across all deployments.

The logging makes use of the AWS Powertools logging lib. This has
simplified our dependencies, and adds some extra fields to our logs for
free. It also handles clean up of the MDC for us.

Instantiation of the loggers no longer uses slf4j. We're not going to
be changing the logging framework anytime soon and it makes things a
little simpler.

This change has also altered the responses of some of the lambdas in
some situations. Rather than returning a journey error response, I've
changed it to just returning a 400. This happens where we the
ipv-session-id has not been sent in the header. If this ever happens
something is seriously wrong and trying to continue the journey is
probably a bad idea and we should just throw. Opinions encouraged.

The passport CRI works differently than the core in that we don't have a
user state table with a unique identifier. Instead, the users state is
managed in the frontend app. As such a bit more work has had to be done
to have a common identifier that we can log across the lambdas.

The two lambdas called by the frontend app are now expecting to receive
a session ID in the headers. There is a PR for the frontend app to
start supplying this. This session ID is then persisted via the auth
code and access token tables so it can be retrieved and logged in the
access token and resource endpoints. Using the session ID from the
frontend app will also help link the lambda invocations to logs from the
frontends.