BAU: Set @Logging annotation and add jackson yaml
This brings in a couple of fixes for the logging in passport.

The previous fix of adding the jackson-dataformat-yaml lib to allow
parsing of the log4j2 config file was undone when I added powertools.
It's still needed so it's being reintroduced.

Secondly, I completely forget to add the quite important `@Logging`
annotation to our handlers. 🤦🏻‍♂️