BAU: Update log field names
Make the log field names camel case to be consistent with what we log in
core. This also changes the session_id field to passportSessionId. This
is to help to differentiate it from other session IDs we may end up
logging, such as persistent session ID. It's also consistent with what
we're logging in core. This will also get logged in the passport front
so we can tie frontend and backend logs together.