BAU: Parse auth grant instead of token request
When receiving an request to the access token endpoint we have been
manually recreating a TokenRequest object from an HTTPRequest and an
arbitrary URL. Rather than jumping through these hoops we can just parse
the auth grant and scope from the params directly.

This also updates the scope returned in the access token to a default
value. The core does not include a scope in the token request. The OAuth
spec states in this scenario the auth server should return a default.