PYIC-1555: Conditionally use resource ID from session
The issuecredential lambda should check to see if the token enty it is
using contains a dcs resource id. If it does, it should be used to to
generate the VC. If not, the session id should be used to open the
session object and use the dcs resource id from there, genearting the VC
as normal.