PYIC-1556: Ensure we can create access tokens with null resource ID
When performing the token exchange the authcode table has an entry for
the dcs response resource id. Later changes will mean that this value
will be blank. Ensure that the code can accommodate the value being
blank, creating a token entry with a blank dcs resource id entry in that
case.

This adds a test that shows there is no issue with the value being
persisted being null