PYIC-1594 Fix Deployment of New Lambdas
The `AWS::Serverless::Function` will create a log group if one does not
exist with the expected log group name. We define our log groups in the
template rather than relying on the serverless transform because we want
to set the retention period. The initial PR introduced two new functions
and their log groups, the log group definition referenced the function
which meant the function would be created first, causing a "log group
with the same name exists" error when eventually CFN came to create the
log group resource. This change adds a depends on between the function
and the log group so the log group should be created first.