PYIC-1594 Remove Provisioned Concurrency
Passport back stack is failing to deploy into the staging environment
whilst timing out attempting to add the provisioned concurrency to the
new version before re-weighting the alias to the new version; it also
times out when trying to roll back. This change temporarily removed
provisioned concurrency to see if the stack will successfully deploy. We
can then work to reinstate it.